@@ -4,7 +4,7 @@ version: 1.0.1
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.24.3"
+  flutter: ">=3.32.0"
 
 dependencies:
   flutter:
